refactor: split KYC into platform-specific components with shared logic
- Replace kyc.tsx with kyc.web.tsx (Didit SDK widget + iframe fallback)
- Replace card/kyc_mobile.tsx with kyc.native.tsx (native SDK + browser fallback)
- Extract shared logic into components/kyc/:
  - useDiditSession: session creation, status polling, state management
  - KycStatusViews: loading, error, completed, native-waiting UI components
- Add @didit-protocol/sdk-web and @didit-protocol/sdk-react-native packages
- Metro/Expo Router auto-resolves .web.tsx and .native.tsx per platform

(cherry picked from commit b32cd2c57a69c8558b5bba5db5701676ce018d01)
