+++ components/kyc/KycStatusViews.tsx
@@ -8,12 +8,20 @@ export function KycLoading() {
   return (
     <View className="flex-1 items-center justify-center py-20">
       <ActivityIndicator size="large" color="#94F27F" />
-      <Text className="mt-4 text-center text-[#ACACAC]">Preparing verification...</Text>
+      <Text className="mt-4 text-center text-[#ACACAC]">
+        Preparing verification...
+      </Text>
     </View>
   );
 }
 
-export function KycError({ message, onRetry }: { message: string; onRetry: () => void }) {
+export function KycError({
+  message,
+  onRetry,
+}: {
+  message: string;
+  onRetry: () => void;
+}) {
   return (
     <View className="flex-1 items-center justify-center gap-4 py-20">
       <Text className="text-center text-red-400">{message}</Text>
