+++ components/kyc/useDiditSession.ts
@@ -1,14 +1,13 @@
 import { useCallback, useEffect, useRef, useState } from 'react';
-import Toast from 'react-native-toast-message';
 import { useRouter } from 'expo-router';
 import { useQueryClient } from '@tanstack/react-query';
+import Toast from 'react-native-toast-message';
 
-import { path } from '@/constants/path';
 import { TRACKING_EVENTS } from '@/constants/tracking-events';
 import { CARD_STATUS_QUERY_KEY } from '@/hooks/useCardStatus';
 import { track } from '@/lib/analytics';
-import { createDiditSession, getCardStatus, getDiditVerificationStatus } from '@/lib/api';
-import { KycStatus, RainApplicationStatus } from '@/lib/types';
+import { createDiditSession, getDiditVerificationStatus } from '@/lib/api';
+import { path } from '@/constants/path';
 import { withRefreshToken } from '@/lib/utils';
 
 export type SessionState =
@@ -33,30 +32,16 @@ export function useDiditSession() {
     try {
       track(TRACKING_EVENTS.KYC_LINK_PAGE_LOADED, { mode: 'didit' });
       const res = await withRefreshToken(() => createDiditSession());
-      if (!res) {
-        setSession({
-          phase: 'error',
-          message: 'Failed to create verification session',
-        });
-        return;
-      }
-      const verificationUrl = res.verification_url ?? res.url;
-      if (!verificationUrl) {
-        setSession({
-          phase: 'error',
-          message: 'No verification URL in session response',
-        });
-        return;
-      }
+      if (!res) return;
       setSession({
         phase: 'ready',
-        verificationUrl,
+        verificationUrl: res.verification_url,
         sessionToken: res.session_token,
       });
     } catch (e: any) {
       const message = e?.message || 'Failed to create verification session';
       setSession({ phase: 'error', message });
-      Toast.show({ type: 'error', text1: 'Error', text2: message, props: { badgeText: '' } });
+      Toast.show({ type: 'error', text1: 'Error', text2: message });
     }
   }, []);
 
@@ -64,77 +49,47 @@ export function useDiditSession() {
     setSession({ phase: 'started' });
   }, []);
 
-  const redirectBasedOnKycStatus = useCallback(
-    async (kycStatus: KycStatus) => {
-      setSession({ phase: 'completed' });
-      queryClient.invalidateQueries({ queryKey: [CARD_STATUS_QUERY_KEY] });
-
-      if (kycStatus === KycStatus.APPROVED) {
-        // KYC approved: check Rain application status for correct redirect
-        try {
-          const cardStatusResponse = await withRefreshToken(() => getCardStatus());
-          if (cardStatusResponse?.rainApplicationStatus === RainApplicationStatus.APPROVED) {
-            router.replace(path.CARD_READY as any);
-            return;
-          }
-        } catch {
-          // Fall through to ready page on approved KYC
-        }
-        router.replace(path.CARD_READY as any);
-      } else if (kycStatus === KycStatus.UNDER_REVIEW) {
-        router.replace(path.CARD_PENDING as any);
-      } else {
-        router.replace(`${String(path.CARD_ACTIVATE)}?kycStatus=${kycStatus}` as any);
-      }
-    },
-    [queryClient, router],
-  );
-
-  const onVerificationComplete = useCallback(() => {
-    Toast.show({
-      type: 'success',
-      text1: 'Verification complete',
-      text2: 'Your identity has been verified.',
-      props: { badgeText: '' },
-    });
-    redirectBasedOnKycStatus(KycStatus.APPROVED);
-  }, [redirectBasedOnKycStatus]);
-
-  const onVerificationPending = useCallback(() => {
-    Toast.show({
-      type: 'info',
-      text1: 'Verification submitted',
-      text2: 'Your verification is being processed.',
-      props: { badgeText: '' },
-    });
-    redirectBasedOnKycStatus(KycStatus.UNDER_REVIEW);
-  }, [redirectBasedOnKycStatus]);
-
-  const onVerificationError = useCallback((message: string) => {
-    Toast.show({
-      type: 'error',
-      text1: 'Verification failed',
-      text2: message,
-      props: { badgeText: '' },
-    });
-    setSession({ phase: 'error', message });
-  }, []);
-
-  // Poll for verification status while SDK is active
+  // Poll for verification status after SDK started
   useEffect(() => {
     if (session.phase !== 'started') return;
 
     const interval = setInterval(async () => {
       try {
-        const status = await withRefreshToken(() => getDiditVerificationStatus());
+        const status = await withRefreshToken(() =>
+          getDiditVerificationStatus(),
+        );
         if (!status) return;
 
-        if (status.status === 'Approved' || status.kycStatus === 'approved') {
+        if (
+          status.status === 'Approved' ||
+          status.kycStatus === 'approved'
+        ) {
           clearInterval(interval);
-          onVerificationComplete();
-        } else if (status.status === 'Declined' || status.kycStatus === 'rejected') {
+          setSession({ phase: 'completed' });
+          queryClient.invalidateQueries({
+            queryKey: [CARD_STATUS_QUERY_KEY],
+          });
+          Toast.show({
+            type: 'success',
+            text1: 'Verification complete',
+            text2: 'Your identity has been verified.',
+          });
+          router.replace(String(path.CARD_ACTIVATE) as any);
+        } else if (
+          status.status === 'Declined' ||
+          status.kycStatus === 'rejected'
+        ) {
           clearInterval(interval);
-          onVerificationError('Your identity verification was declined. Please try again.');
+          Toast.show({
+            type: 'error',
+            text1: 'Verification failed',
+            text2:
+              'Your identity verification was declined. Please try again.',
+          });
+          setSession({
+            phase: 'error',
+            message: 'Verification declined',
+          });
         }
       } catch {
         // silently retry on network errors
@@ -142,7 +97,7 @@ export function useDiditSession() {
     }, POLL_INTERVAL_MS);
 
     return () => clearInterval(interval);
-  }, [session.phase, onVerificationComplete, onVerificationError]);
+  }, [session.phase, queryClient, router]);
 
   // Auto-init on mount
   useEffect(() => {
@@ -151,10 +106,8 @@ export function useDiditSession() {
 
   return {
     session,
+    sdkInitializedRef,
     initSession,
     markStarted,
-    onVerificationComplete,
-    onVerificationPending,
-    onVerificationError,
   };
 }
