+++ package.json
@@ -36,6 +36,8 @@
     "@apollo/client": "^4.1.6",
     "@braintree/browser-detection": "^2.1.1",
     "@coinbase/wallet-mobile-sdk": "^1.1.2",
+    "@didit-protocol/sdk-react-native": "^1.0.0",
+    "@didit-protocol/sdk-web": "^1.0.0",
     "@cryptoalgebra/fuse-sdk": "npm:@cryptoalgebra/fuse-sdk@latest",
     "@didit-protocol/sdk-react-native": "^3.2.5",
     "@didit-protocol/sdk-web": "^0.1.8",
